:recycle: 테스트 결과 페이지 동적 라우팅 적용
