+++ src/api/test/test.ts
@@ -1,11 +1,11 @@
 import { post } from '@/lib/axios';
 import { SuccessResponse } from '@/types/response';
-import { TestFormType } from '@/types/test';
+import { TestFormResponseType, TestFormType } from '@/types/test';
 
 //api
 
 const postTestResult = async (data: TestFormType) => {
-  const {data : result} = await post<SuccessResponse<TestFormType>>('/test/result', data);
+  const {data : result} = await post<SuccessResponse<TestFormResponseType>>('/test/result', data);
 
   return result;
 };
