+++ src/app/test/result/[id]/_components/ImageBox.tsx
