+++ src/app/test/result/[id]/_components/ResultContents.tsx
