+++ src/app/test/result/[id]/_components/ShareBox.tsx
