+++ src/app/test/result/[id]/_components/TempertaureBox.tsx
