+++ src/app/test/result/[id]/_components/index.ts
