+++ src/hooks/api/test/useCreateTestResult.tsx
@@ -8,7 +8,7 @@ export const useCreateTestResult = () => {
     mutationFn: (data: TestFormType) => TEST.POST_RESULT(data),
     //TODO : 성공 및 에러 처리
     onSuccess: (response) => {
-      console.log('Post 성공', response);
+      console.log('Post 성공, id값', response.data.id);
     },
     onError: (error) => {
       console.log('Post 실패', error);
