--- src/types/test/test.ts
@@ -1,8 +1,0 @@
-export type TestFormType = {
-  buddy: string;
-  gender: string;
-  age: string;
-  trust: number;
-  love: number;
-  talk: number;
-};
