+++ src/types/test/test.tsx
@@ -1,0 +1,17 @@
+export type TestFormType = {
+  buddy: string;
+  gender: string;
+  age: string;
+  trust: number;
+  love: number;
+  talk: number;
+};
+
+export type TestFormResponseType = TestFormType & {
+  id: number;
+  temperature: number;
+  imageUrl: string;
+  description: string;
+  title: string;
+  createdAt: string;
+};
